# openapi.yaml
openapi: 3.1.0
info:
  title: Zantara Bridge API
  description: |
    Comprehensive API for Google Workspace integration, AI-powered chat, and productivity tools.
    The API provides secure endpoints for Drive, Gmail, Calendar operations, AI analytics, and conversation management.
    
    **Performance Features:**
    - Redis-based caching for Drive operations (5min folder cache, 10min file cache, 3min search cache)
    - Firestore-powered memory system with full-text search and relevance scoring
    - Automatic cache invalidation on data changes
    - Zero-cost optimized architecture with graceful fallbacks
  version: 1.0.0
  contact:
    name: Bali Zero HQ
    email: zero@balizero.com
  license:
    name: Proprietary
servers:
  - url: https://zantara-bridge-v2-prod-1064094238013.asia-southeast2.run.app
    description: Production server (Cloud Run)
  - url: https://zantara-light-bridge-1064094238013.asia-southeast2.run.app
    description: Zantara Light Bridge - Zero-cost optimized (Cloud Run)
  - url: http://localhost:8080
    description: Development server
security:
  - ApiKeyAuth: []
  - UserIdentification: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key for authentication (required for most endpoints)
    UserIdentification:
      type: apiKey
      in: header
      name: X-BZ-USER
      description: User identification header (defaults to 'BOSS' if not provided)
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
      required:
        - error
...
    ChatRequest:
      type: object
      properties:
        message:
          type: string
          description: The chat message to process
        ririMode:
          type: boolean
          description: Enable RIRI mode for special processing
        sessionId:
          type: string
          description: Session ID for conversation continuity
        saveAs:
          type: string
          enum: [chat, note, brief]
          description: How to save the conversation
        title:
          type: string
          description: Title for saved conversation
      required:
        - message
    ChatResponse:
      type: object
      properties:
        ok:
          type: boolean
        text:
          type: string
          description: AI response text
        responseTime:
          type: number
          description: Response time in milliseconds
        savedAs:
          type: string
          description: How the conversation was saved
        owner:
          type: string
          description: Owner of the conversation
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
    MonitoringStatus:
      type: object
      properties:
        ok:
          type: boolean
        monitoring:
          type: object
          properties:
            health:
              type: object
            stats:
              type: object
            drive:
              type: object
              properties:
                ok:
                  type: boolean
                user:
                  type: object
                  properties:
                    displayName:
                      type: string
                    emailAddress:
                      type: string
                quota:
                  type: object
                isImpersonated:
                  type: boolean
                impersonateUser:
                  type: string
            openai:
              type: object
            gmail:
              type: object
            calendar:
              type: object
        ts:
          type: string
          format: date-time
paths:
...
  /api/monitoring:
    get:
      tags: [Monitoring]
      summary: Comprehensive monitoring status
      description: Returns detailed monitoring information including Drive impersonation status
      responses:
        '200':
          description: Monitoring status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/chat:
    post:
      tags: [AI & Chat]
      summary: Main chat endpoint
      description: Process chat messages with AI and optionally save to Drive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /actions/memory/save:
    post:
      tags: [Memory]
      summary: Save memory entry (Firestore Implementation)
      description: |
        Save a memory entry to Firestore with full metadata support.
        Supports tags, categories, relevance scoring, and automatic access tracking.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: User identifier for memory isolation
                  example: "user-123"
                content:
                  type: string
                  description: Main memory content
                  example: "Meeting notes about quarterly planning"
                title:
                  type: string
                  description: Optional memory title
                  example: "Q4 Planning Meeting"
                tags:
                  type: array
                  items:
                    type: string
                  description: Searchable tags
                  example: ["meeting", "planning", "q4"]
                category:
                  type: string
                  description: Memory category
                  example: "work"
                source:
                  type: string
                  description: Origin of the memory
                  example: "api"
                metadata:
                  type: object
                  properties:
                    tokenCount:
                      type: number
                    contextType:
                      type: string
                      enum: [conversation, document, note, task]
                  description: Additional metadata
              required:
                - userId
                - content
      responses:
        '200':
          description: Memory entry saved successfully to Firestore
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Firestore document ID
                        example: "abc123def456"
                      message:
                        type: string
                        example: "Memory entry saved successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/users:
    get:
      tags: [User Management]
      summary: List all users
      description: Retrieve a list of all registered users
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserProfile'
                  total:
                    type: integer
    
    post:
      tags: [User Management]
      summary: Register new user
      description: Register a new user and generate API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
              required:
                - username
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  apiKey:
                    type: string
                  username:
                    type: string
                  message:
                    type: string
  /api/conversations:
    get:
      tags: [Conversations]
      summary: List all conversations
      description: Retrieve all conversations with filtering options
      parameters:
        - name: collab
          in: query
          description: Filter by collaborator
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, closed, archived]
      responses:
        '200':
          description: Conversations retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks
      description: Retrieve all configured webhooks
      responses:
        '200':
          description: Webhooks retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        url:
                          type: string
                        events:
                          type: array
                          items:
                            type: string
                        active:
                          type: boolean
                        created:
                          type: string
                          format: date-time
                  total:
                    type: integer
    
    post:
      tags: [Webhooks]
      summary: Create webhook
      description: Register a new webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
              required:
                - url
                - events
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  url:
                    type: string
                  events:
                    type: array
                    items:
                      type: string
                  secret:
                    type: string
                  active:
                    type: boolean
                  created:
                    type: string
                    format: date-time
...